# docs/api/ledger.openapi.json
{
  "components" : {
    "schemas" : {
      "Money" : {
        "properties" : {
          "asset" : {
            "type" : "string"
          },
          "atomicUnits" : {
            "format" : "int64",
            "type" : "integer"
          },
          "decimals" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "asset", "atomicUnits", "decimals" ],
        "type" : "object"
      },
      "PaymentDetail" : {
        "properties" : {
          "entries" : {
            "items" : {
              "$ref" : "#/components/schemas/PaymentEntry"
            },
            "type" : "array"
          },
          "mismatches" : {
            "items" : {
              "$ref" : "#/components/schemas/PaymentFinding"
            },
            "type" : "array"
          },
          "payment" : {
            "$ref" : "#/components/schemas/PaymentSummary"
          }
        },
        "required" : [ "entries", "mismatches", "payment" ],
        "type" : "object"
      },
      "PaymentEntry" : {
        "properties" : {
          "book" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "effectiveAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "entryId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "postings" : {
            "items" : {
              "$ref" : "#/components/schemas/PaymentPosting"
            },
            "type" : "array"
          },
          "reversesEntryId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "book", "description", "effectiveAt", "entryId", "kind", "postings" ],
        "type" : "object"
      },
      "PaymentFinding" : {
        "properties" : {
          "adjustmentEntryId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "chainValue" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/Money"
            }, {
              "type" : "null"
            } ]
          },
          "detectedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "ledgerValue" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/Money"
            }, {
              "type" : "null"
            } ]
          },
          "reconciliationRunId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "string"
          }
        },
        "required" : [ "detectedAt", "kind", "status" ],
        "type" : "object"
      },
      "PaymentPage" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/PaymentSummary"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "items" ],
        "type" : "object"
      },
      "PaymentPosting" : {
        "properties" : {
          "accountCode" : {
            "type" : "string"
          },
          "amount" : {
            "$ref" : "#/components/schemas/Money"
          },
          "side" : {
            "type" : "string"
          }
        },
        "required" : [ "accountCode", "amount", "side" ],
        "type" : "object"
      },
      "PaymentSummary" : {
        "properties" : {
          "amount" : {
            "$ref" : "#/components/schemas/Money"
          },
          "buyerState" : {
            "type" : "string"
          },
          "buyerTxHash" : {
            "type" : [ "string", "null" ]
          },
          "chainState" : {
            "type" : "string"
          },
          "chainTxHash" : {
            "type" : [ "string", "null" ]
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "payTo" : {
            "type" : "string"
          },
          "paymentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "runId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "sellerState" : {
            "type" : "string"
          },
          "sellerTxHash" : {
            "type" : [ "string", "null" ]
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "amount", "buyerState", "chainState", "createdAt", "payTo", "paymentId", "sellerState", "updatedAt" ],
        "type" : "object"
      },
      "ProblemDetail" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "properties" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ReconciliationCounters" : {
        "properties" : {
          "checked" : {
            "format" : "int32",
            "type" : "integer"
          },
          "matched" : {
            "format" : "int32",
            "type" : "integer"
          },
          "mismatches" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pending" : {
            "format" : "int32",
            "type" : "integer"
          },
          "resolvedUnused" : {
            "format" : "int32",
            "type" : "integer"
          },
          "resolvedUsed" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "checked", "matched", "mismatches", "pending", "resolvedUnused", "resolvedUsed" ],
        "type" : "object"
      },
      "ReconciliationItem" : {
        "properties" : {
          "amount" : {
            "$ref" : "#/components/schemas/Money"
          },
          "buyerState" : {
            "type" : "string"
          },
          "chainState" : {
            "type" : "string"
          },
          "mismatch" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/ReconciliationMismatch"
            }, {
              "type" : "null"
            } ]
          },
          "payTo" : {
            "type" : "string"
          },
          "payer" : {
            "type" : "string"
          },
          "paymentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "paymentIntentId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "runId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "sellerState" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "txHash" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "amount", "buyerState", "chainState", "payTo", "payer", "paymentId", "sellerState", "status" ],
        "type" : "object"
      },
      "ReconciliationMismatch" : {
        "properties" : {
          "adjustmentEntryId" : {
            "format" : "uuid",
            "type" : [ "string", "null" ]
          },
          "chainValue" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/Money"
            }, {
              "type" : "null"
            } ]
          },
          "kind" : {
            "type" : "string"
          },
          "ledgerValue" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/Money"
            }, {
              "type" : "null"
            } ]
          }
        },
        "required" : [ "kind" ],
        "type" : "object"
      },
      "ReconciliationReport" : {
        "properties" : {
          "finishedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/ReconciliationItem"
            },
            "type" : "array"
          },
          "network" : {
            "type" : "string"
          },
          "runId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "safeBlock" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "summary" : {
            "$ref" : "#/components/schemas/ReconciliationCounters"
          },
          "suspense" : {
            "$ref" : "#/components/schemas/Money"
          },
          "suspenseSide" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "items", "network", "runId", "startedAt", "status", "summary", "suspense" ],
        "type" : "object"
      },
      "ReconciliationRunList" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/ReconciliationRunSummary"
            },
            "type" : "array"
          }
        },
        "required" : [ "items" ],
        "type" : "object"
      },
      "ReconciliationRunSummary" : {
        "properties" : {
          "finishedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "runId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "safeBlock" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "summary" : {
            "$ref" : "#/components/schemas/ReconciliationCounters"
          }
        },
        "required" : [ "runId", "startedAt", "status", "summary" ],
        "type" : "object"
      },
      "ReconciliationStarted" : {
        "properties" : {
          "runId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "runId" ],
        "type" : "object"
      },
      "RevenueReport" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/SellerRevenue"
            },
            "type" : "array"
          }
        },
        "required" : [ "items" ],
        "type" : "object"
      },
      "SellerRevenue" : {
        "properties" : {
          "creditNotes" : {
            "$ref" : "#/components/schemas/Money"
          },
          "credited" : {
            "format" : "int64",
            "type" : "integer"
          },
          "customerCredits" : {
            "$ref" : "#/components/schemas/Money"
          },
          "grossSales" : {
            "$ref" : "#/components/schemas/Money"
          },
          "netRevenue" : {
            "$ref" : "#/components/schemas/SignedAmount"
          },
          "payTo" : {
            "type" : "string"
          },
          "sales" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "creditNotes", "credited", "customerCredits", "grossSales", "netRevenue", "payTo", "sales" ],
        "type" : "object"
      },
      "SignedAmount" : {
        "properties" : {
          "asset" : {
            "type" : "string"
          },
          "atomicUnits" : {
            "format" : "int64",
            "type" : "integer"
          },
          "decimals" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "asset", "atomicUnits", "decimals" ],
        "type" : "object"
      },
      "TrialBalanceRow" : {
        "properties" : {
          "account" : {
            "type" : "string"
          },
          "asset" : {
            "type" : "string"
          },
          "balance" : {
            "type" : "integer"
          },
          "book" : {
            "type" : "string"
          },
          "credit" : {
            "type" : "integer"
          },
          "debit" : {
            "type" : "integer"
          },
          "decimals" : {
            "format" : "int32",
            "type" : "integer"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "account", "asset", "balance", "book", "credit", "debit", "decimals", "type" ],
        "type" : "object"
      }
    }
  },
  "info" : {
    "description" : "Double-entry ledger, seller revenue and on-chain reconciliation (testnet only).",
    "title" : "Saiman ledger API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/ledger/payments" : {
      "get" : {
        "operationId" : "payments",
        "parameters" : [ {
          "in" : "query",
          "name" : "runId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "book",
          "required" : false,
          "schema" : {
            "enum" : [ "BUYER", "SELLER" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "before",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentPage"
                }
              }
            },
            "description" : "A page of payments"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "Malformed runId, book, limit or before"
          }
        },
        "tags" : [ "ledger-payments-controller" ]
      }
    },
    "/api/v1/ledger/payments/{paymentId}" : {
      "get" : {
        "operationId" : "payment",
        "parameters" : [ {
          "in" : "path",
          "name" : "paymentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentDetail"
                }
              }
            },
            "description" : "The payment"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "paymentId is not a UUID"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "No such payment"
          }
        },
        "tags" : [ "ledger-payments-controller" ]
      }
    },
    "/api/v1/ledger/revenue" : {
      "get" : {
        "operationId" : "revenue",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RevenueReport"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "revenue-controller" ]
      }
    },
    "/api/v1/ledger/trial-balance" : {
      "get" : {
        "operationId" : "trialBalance",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TrialBalanceRow"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "trial-balance-controller" ]
      }
    },
    "/api/v1/reconciliation/runs" : {
      "get" : {
        "operationId" : "runs",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationRunList"
                }
              }
            },
            "description" : "Runs, newest first"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "limit outside 1..100"
          }
        },
        "tags" : [ "reconciliation-controller" ]
      },
      "post" : {
        "operationId" : "start",
        "responses" : {
          "202" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationStarted"
                }
              }
            },
            "description" : "The run started in the background"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "A run is in progress"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "The previous manual run started too recently",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds to wait",
                "schema" : {
                  "type" : "integer"
                },
                "style" : "simple"
              }
            }
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "No Base Sepolia client is configured"
          }
        },
        "tags" : [ "reconciliation-controller" ]
      }
    },
    "/api/v1/reconciliation/runs/latest" : {
      "get" : {
        "operationId" : "latest",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationReport"
                }
              }
            },
            "description" : "The latest run's report"
          },
          "404" : {
            "description" : "No run yet (empty body)"
          }
        },
        "tags" : [ "reconciliation-controller" ]
      }
    },
    "/api/v1/reconciliation/runs/{id}" : {
      "get" : {
        "operationId" : "report",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationReport"
                }
              }
            },
            "description" : "The run's report"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description" : "id is not a UUID"
          },
          "404" : {
            "description" : "No such run (empty body)"
          }
        },
        "tags" : [ "reconciliation-controller" ]
      }
    }
  }
}